Add migration changeaddressv2 for Stores table update
This migration modifies the `Stores` table by dropping the `UserId` column and making the `AppUserId` column non-nullable. It also establishes a foreign key relationship between `AppUserId` in the `Stores` table and the `Id` column in the `AspNetUsers` table with cascade delete behavior. The `Down` method restores the previous state by re-adding the `UserId` columns and making `AppUserId` nullable again.